feat(levels)!: Support custom log levels in LogDNA
As Winston allows for the use of custom log levels, now
so does LogDNA. Custom levels can be defined for both
Winston and LogDNA such that those custom levels will
show up on LogDNA's log viewer.

BREAKING CHANGE: This removes the log level "translation"
that used to be in place to convert Winston levels to ones
that would be acceptable by LogDNA. Since LogDNA can now
define custom levels as well, this translation is no longer
needed, however it may break implementations that are
relying on the translated levels.

Fixes: #31