fix(levels): Automatically set up LogDNA custom levels for defaults
Winston uses "levels" as an object where the values are numeric.
If no levels are specified, it uses the "npm" levels that are
defined in their docs. These levels are not all supported by
LogDNA, so if "levels" is used, automatically set up custom
LogDNA levels based on the keys of the winston levels object.