fix(level): Separate concerns for the "level" parameter
Both Winston and LogDNA have a "level" parameter, but they
mean different things. In Winston, it represents a "max"
level wherein anything logged above that level's rank will
not be logged (and thus sent to LogDNA). To disambiguate
this logger transport, the parameter will be called `maxLevel`,
and will only apply to Winston. LogDNA's `level` parameter is
actually not usable with the implementation of this transport
since all logs will be specifying a level value.

Fixes: #33